fix: Can not review user.

+++ backend2/src/main/java/com/cmoney/backend2/forumocean/service/api/rating/RatingComment.kt
@@ -13,5 +13,5 @@ data class RatingComment(
     @SerializedName("comment")
     val comment: String?,
     @SerializedName("score")
-    val score: Double?
+    val score: Int?
 )
